break circular deps: file extraction, barrel bypass, class co-location
