+++ packages/ketcher-core/src/application/editor/operations/rgroupAttachmentPoint/RGroupAttachmentPointAdd.ts
@@ -4,7 +4,6 @@ import {
   RGroupAttachmentPoint,
   RGroupAttachmentPointType,
 } from 'domain/entities';
-import { RGroupAttachmentPointRemove } from './RGroupAttachmentPointRemove';
 import { OperationPriority, OperationType } from '../OperationType';
 import BaseOperation from '../BaseOperation';
 
@@ -60,4 +59,52 @@ class RGroupAttachmentPointAdd extends BaseOperation {
   }
 }
 
-export { RGroupAttachmentPointAdd };
+type RemoveData = {
+  atomId: number;
+  attachmentPointType;
+  attachmentPointId: number;
+};
+
+const INITIAL_DATA = {
+  atomId: 0,
+  attachmentPointType: 1,
+  attachmentPointId: 0,
+};
+
+class RGroupAttachmentPointRemove extends BaseOperation {
+  readonly data: RemoveData;
+
+  constructor(attachmentPointId: number) {
+    super(
+      OperationType.R_GROUP_ATTACHMENT_POINT_REMOVE,
+      OperationPriority.R_GROUP_ATTACHMENT_POINT_REMOVE,
+    );
+    this.data = {
+      ...INITIAL_DATA,
+      attachmentPointId,
+    };
+  }
+
+  execute(restruct: ReStruct) {
+    const { attachmentPointId } = this.data;
+    const struct = restruct.molecule;
+    const item = struct.rgroupAttachmentPoints.get(attachmentPointId);
+    const reItem = restruct.rgroupAttachmentPoints.get(attachmentPointId);
+    assert(item != null && reItem != null);
+
+    this.data.atomId = item.atomId;
+    this.data.attachmentPointType = item.type;
+
+    restruct.markItemRemoved();
+    restruct.clearVisel(reItem.visel);
+    restruct.rgroupAttachmentPoints.delete(attachmentPointId);
+
+    struct.rgroupAttachmentPoints.delete(attachmentPointId);
+  }
+
+  invert() {
+    return new RGroupAttachmentPointAdd(this.data);
+  }
+}
+
+export { RGroupAttachmentPointAdd, RGroupAttachmentPointRemove };
